+++ packages/kernel/src/Actor.js
@@ -58,20 +58,31 @@ export class Actor extends WorldcoreModel {
         super.destroy();
     }
 
-    set(options = {}, noSnap) {
+    set(options = {}) {
         const sorted = Object.entries(options).sort((a,b) => { return b[0] < a[0] ? 1 : -1 } );
         for (const option of sorted) {
             const name = option[0];
             const value = option[1];
+            const ul = "_" + name;
+            const nameSet = name+'Set';
+            if (this[nameSet]) {
+                this[nameSet](value)
+            } else {
+                this[ul] = value;
+            }
+            this.say(nameSet, value);
+        }
+        return sorted;
+    }
 
-            const old = this["_" + name];
-
-            if (this[name+'Set']) this[name+'Set'](value, noSnap)
-
-            this["_" + name] = value;
-            const data = {v: value, o: old};
-            this.say(name + "Set", data);
-            if (!noSnap) this.say(name + "Snap", data);
+    snap(options = {}) {
+        const sorted = this.set(options);
+        for (const option of sorted) {
+            const name = option[0];
+            const value = option[1];
+            const nameSnap = name+'Snap';
+            if (this[nameSnap]) this[nameSnap](value)
+            this.say(nameSnap, value);
         }
     }